KAFKA-4651 updates per comments, remove redundant tests, unused methods
